+++ src/actions/help.js
@@ -1,12 +1,12 @@
 import chalk from 'chalk'
 
-import * as errors from '../constants/errors'
-import * as info from '../constants/info'
-import * as string from '../helpers/string'
+import * as errorHelper from '../helpers/error'
+import * as infoHelper from '../helpers/info'
+import * as stringHelper from '../helpers/string'
 
 export async function help(options, state) {
-  if (state === errors.INVALID_ACTION) {
-    errors.send(state, { action: options.action })
+  if (state === errorHelper.INVALID_ACTION) {
+    errorHelper.send(state, { action: options.action })
     console.log(
       chalk.blue('INFO!'),
       `Try to run ${chalk.magenta('cth --help')} for help`
@@ -14,10 +14,11 @@ export async function help(options, state) {
     return
   }
 
-  if (info[options.action]) {
-    console.log(string.trim(info[options.action]))
-    if (options.action !== 'flags') console.log(string.trim(info.flags))
+  if (infoHelper[options.action]) {
+    console.log(stringHelper.trim(infoHelper[options.action]))
+    if (options.action !== 'flags')
+      console.log(stringHelper.trim(infoHelper.flags))
   } else {
-    console.log(string.trim(info.complete))
+    console.log(stringHelper.trim(infoHelper.complete))
   }
 }