+++ src/actions/info.js
@@ -1,10 +1,9 @@
 import fs from 'fs-extra'
 import chalk from 'chalk'
-import inquirer from 'inquirer'
 
-import * as errors from '../constants/errors'
-import * as paths from '../constants/paths'
-import * as file from '../helpers/file'
+import * as errorHelper from '../helpers/error'
+import * as pathHelper from '../helpers/path'
+import * as fileHelper from '../helpers/file'
 
 export async function info(options) {
   options.templateNames.forEach((templateName) => {
@@ -13,38 +12,24 @@ export async function info(options) {
 }
 
 async function logInfo(options) {
-  const storageData = fs.readJSONSync(paths.DATA_JSON)
+  try {
+    const storageData = fs.readJSONSync(pathHelper.DATA_JSON)
 
-  // prompt question if no template name
-  options = Object.assign(
-    options,
-    await inquirer.prompt([
-      {
-        type: 'input',
-        name: 'templateName',
-        when: !options.templateName,
-        message: 'Please enter template name',
-      },
-    ])
-  )
+    const templatePath = storageData[options.templateName]
+    if (!templatePath) {
+      errorHelper.send(errorHelper.TEMPLATE_NOT_FOUND, {
+        name: options.templateName,
+      })
+      return
+    }
 
-  if (!options.templateName) {
-    errors.send(errors.NO_TEMPLATE_NAME)
-    return
+    const templateDescription =
+      fileHelper.getTemplateData(templatePath).description || ''
+
+    console.log(
+      `${chalk.magenta(options.templateName)}\n${templateDescription}\n`
+    )
+  } catch (err) {
+    errorHelper.send(err)
   }
-
-  const templatePath = storageData[options.templateName]
-  if (!templatePath) {
-    errors.send(errors.TEMPLATE_NOT_FOUND, {
-      name: options.templateName,
-    })
-    return
-  }
-
-  const templateDescription =
-    file.getTemplateData(templatePath).description || ''
-
-  console.log(
-    `${chalk.magenta(options.templateName)}\n${templateDescription}\n`
-  )
 }