+++ src/actions/list.js
@@ -1,17 +1,23 @@
 import fs from 'fs-extra'
 import chalk from 'chalk'
 
-import * as paths from '../constants/paths'
-import * as file from '../helpers/file'
+import * as pathHelper from '../helpers/path'
+import * as errorHelper from '../helpers/error'
+import * as fileHelper from '../helpers/file'
 
 export async function list() {
-  const storageData = fs.readJSONSync(paths.DATA_JSON)
+  try {
+    const storageData = fs.readJSONSync(pathHelper.DATA_JSON)
 
-  let string = ''
-  Object.keys(storageData).forEach((key) => {
-    const desc = file.getTemplateData(storageData[key]).description || '...'
-    string += `- ${chalk.magenta(key)}\n${desc}\n\n`
-  })
+    let string = ''
+    Object.keys(storageData).forEach((key) => {
+      const desc =
+        fileHelper.getTemplateData(storageData[key]).description || '...'
+      string += `- ${chalk.magenta(key)}\n${desc}\n\n`
+    })
 
-  console.log(string.replace(/\n\n$/, ''))
+    console.log(string.replace(/\n\n$/, ''))
+  } catch (err) {
+    errorHelper.send(err)
+  }
 }