+++ src/helpers/error.js
@@ -1,4 +1,4 @@
-import * as chalk from 'chalk'
+import chalk from 'chalk'
 
 export const INVALID_ACTION = new Error('Invalid action $action!')
 export const NO_TEMPLATE_JSON = new Error(
@@ -8,13 +8,17 @@ export const NO_TEMPLATE_NAME = new Error(
   'Please provide a unique template name for this template!'
 )
 export const NO_BACKUP_DATA = new Error(
-  'Backup folder must contain backup.json file'
+  'Backup folder must contain backup.json file!'
 )
 export const TEMPLATE_EXIST = new Error(
   'Template with name $name already exist!'
 )
 export const TEMPLATE_NOT_FOUND = new Error(
   'Template with name $name not found!'
+)
+export const GIT_NOT_FOUND = new Error('This action need to executed git!')
+export const INVALID_TEMPLATE_NAME = new Error(
+  'A template name cannot started with "github@" "gitlab@" or "bitbucket@"'
 )
 
 export const format = (message) => `${chalk.red.bold('ERR!')} ${message}`
@@ -27,5 +31,5 @@ export const send = (err, params) => {
   })
 
   console.log(format(message))
-  return false
+  throw err
 }