+++ src/helpers/info.js
@@ -1,6 +1,6 @@
 import chalk from 'chalk'
 import fs from 'fs-extra'
-import * as paths from './paths'
+import * as paths from './path'
 
 const VERSION = fs.readJSONSync(paths.PACKAGE_JSON).version
 
@@ -18,11 +18,11 @@ export const save = `
 `
 export const clone = `
    ${chalk.blue.bold('Clone a template:')}
-   |  ${chalk.yellow('cth clone ...[template-name]')}
+   |  ${chalk.yellow('cth clone ...[template-name or git-repo]')}
 `
 export const update = `
    ${chalk.blue.bold('Update added template:')}
-   | ${chalk.yellow('cth update [template-name] [path/to/template/directory]')}
+   | ${chalk.yellow('cth update [template-name]')}
 `
 export const backup = `
    ${chalk.blue.bold('Backup your local template storage:')}
@@ -49,18 +49,27 @@ export const list = `
    |  ${chalk.yellow('cth list')}
 `
 
+export const rename = `
+   ${chalk.blue.bold('Rename saved template:')}
+   |  ${chalk.yellow('cth rename [template-name] [new-tmplate-name]')}
+`
+
 export const flags = `
    |
    |  flags:
    |     ${chalk.yellow('-h, --help')} : get command's information
    |     ${chalk.yellow('-v, --version')} : get cithak version
-   |     ${chalk.yellow('-p, --path')} : add extra path relative to current working directory ${chalk.gray('( ex: -p ./my-folder )')}
+   |     ${chalk.yellow('-p, --path')} : add extra path relative to current working directory or a git repo
+   |     ${chalk.gray('( ex: -p ./my-folder or -p github@username/repo )')}
+
    |     ${chalk.yellow('-y, --yes')} : exxecute all template tasks
    |     ${chalk.yellow('-s, --safe')} : copy template without replacing existing files
-   |     ${chalk.yellow('-c, --change')} : change file that will be executed ${chalk.gray('( ex: \'npm|pnpm\' ) : npm i package -> pnpm i package')}
+   |     ${chalk.yellow('-c, --change')} : change file that will be executed
+   |     ${chalk.gray('( ex: \'npm|pnpm\' ) : npm i package -> pnpm i package')}
+
    |     ${chalk.yellow('--no-join')} : copy template without join joinable data file like json file
    |     ${chalk.yellow('--no-exec')} : copy template without execute any tasks
 `
 
-export const complete = 
-   main + save + info + list + clone + update + remove + merge + flags
+export const complete =
+  main + save + info + list + clone + update + remove + merge + rename + flags